# String.StartsWith performance - OrdinalCompareSubstring