@bbowyersmyth 2016-01-23
For the curious here is how OrdinalCompareSubstring performs in place of EqualsHelper within `String.Equals(string, string)`.

Two 15 character strings that differ at index n:

     Method | n |   AvrTime |    StdDev |           op/s |
----------- |---------: |----------: |----------: |---------------: |
  EqualsHelper |        0 | 3.5466 ns | 0.1022 ns | 282,112,499.76 |
 OrdinalCompareSubstring |        0 | 3.9923 ns | 0.0174 ns | 250,487,815.52 |
  EqualsHelper |        1 | 3.6788 ns | 0.0119 ns | 271,827,237.50 |
 OrdinalCompareSubstring |        1 | 4.0442 ns | 0.1041 ns | 247,374,159.09 |
  EqualsHelper |        2 | 3.7353 ns | 0.0652 ns | 267,772,805.88 |
 OrdinalCompareSubstring |        2 | 3.9615 ns | 0.0028 ns | 252,430,975.66 |
  EqualsHelper |        3 | 3.7838 ns | 0.0706 ns | 264,343,651.65 |
 OrdinalCompareSubstring |        3 | 3.9750 ns | 0.0144 ns | 251,576,269.63 |
  EqualsHelper |        4 | 3.9790 ns | 0.0515 ns | 251,345,534.61 |
 OrdinalCompareSubstring |        4 | 4.2554 ns | 0.0297 ns | 235,001,013.93 |
  EqualsHelper |        5 | 3.9760 ns | 0.0161 ns | 251,513,954.20 |
 OrdinalCompareSubstring |        5 | 4.2620 ns | 0.0255 ns | 234,637,924.74 |
  EqualsHelper |        6 | 4.0342 ns | 0.1067 ns | 247,998,018.76 |
 OrdinalCompareSubstring |        6 | 4.2443 ns | 0.0309 ns | 235,618,264.70 |
  EqualsHelper |        7 | 3.9775 ns | 0.0223 ns | 251,419,057.30 |
 OrdinalCompareSubstring |        7 | 4.2958 ns | 0.0363 ns | 232,798,663.03 |
  EqualsHelper |        8 | 4.8275 ns | 0.0132 ns | 207,149,689.60 |
 OrdinalCompareSubstring |        8 | 4.2687 ns | 0.0298 ns | 234,271,059.72 |
  EqualsHelper |        9 | 4.8124 ns | 0.0408 ns | 207,805,627.38 |
 OrdinalCompareSubstring |        9 | 4.6896 ns | 0.0710 ns | 213,269,012.35 |
  EqualsHelper |       10 | 4.9161 ns | 0.0634 ns | 203,436,995.32 |
 OrdinalCompareSubstring |       10 | 4.6906 ns | 0.1279 ns | 213,299,091.99 |
  EqualsHelper |       11 | 4.8353 ns | 0.0242 ns | 206,816,261.57 |
 OrdinalCompareSubstring |       11 | 4.3410 ns | 0.1117 ns | 230,462,888.03 |
  EqualsHelper |       12 | 6.3871 ns | 0.1825 ns | 156,650,082.26 |
 OrdinalCompareSubstring |       12 | 4.5727 ns | 0.0438 ns | 218,700,185.46 |
  EqualsHelper |       13 | 6.2458 ns | 0.0328 ns | 160,110,008.34 |
 OrdinalCompareSubstring |       13 | 4.5889 ns | 0.0646 ns | 217,943,991.00 |
  EqualsHelper |       14 | 7.1077 ns | 0.0470 ns | 140,697,135.33 |
 OrdinalCompareSubstring |       14 | 4.5693 ns | 0.0087 ns | 218,851,136.38 |

Two equal length character strings:

     Method | Len |   AvrTime |    StdDev |           op/s |
----------- |---------: |----------: |----------: |---------------: |
  EqualsHelper |        1 | 4.6369 ns | 0.1794 ns | 215,870,166.73 |
 OrdinalCompareSubstring |        1 | 4.5604 ns | 0.0094 ns | 219,278,035.75 |
  EqualsHelper |        2 | 4.5419 ns | 0.0237 ns | 220,177,697.69 |
 OrdinalCompareSubstring |        2 | 4.3023 ns | 0.0429 ns | 232,451,507.11 |
  EqualsHelper |        3 | 5.4157 ns | 0.0225 ns | 184,652,105.07 |
 OrdinalCompareSubstring |        3 | 4.2586 ns | 0.0295 ns | 234,829,099.21 |
  EqualsHelper |        4 | 5.4452 ns | 0.0262 ns | 183,649,890.96 |
 OrdinalCompareSubstring |        4 | 5.1252 ns | 0.0100 ns | 195,116,510.97 |
  EqualsHelper |        5 | 6.2600 ns | 0.0173 ns | 159,746,206.39 |
 OrdinalCompareSubstring |        5 | 5.1961 ns | 0.1305 ns | 192,530,192.50 |
  EqualsHelper |        6 | 6.2754 ns | 0.0456 ns | 159,359,239.59 |
 OrdinalCompareSubstring |        6 | 5.9853 ns | 0.0589 ns | 167,087,349.82 |
  EqualsHelper |        7 | 7.0796 ns | 0.0758 ns | 141,261,075.88 |
 OrdinalCompareSubstring |        7 | 5.9765 ns | 0.0278 ns | 167,323,456.00 |
  EqualsHelper |        8 | 7.1740 ns | 0.1392 ns | 139,427,727.60 |
 OrdinalCompareSubstring |        8 | 7.0893 ns | 0.1048 ns | 141,078,981.19 |
  EqualsHelper |        9 | 8.0171 ns | 0.0145 ns | 124,734,229.93 |
 OrdinalCompareSubstring |        9 | 6.8691 ns | 0.0456 ns | 145,583,918.13 |
  EqualsHelper |       10 | 8.0684 ns | 0.0516 ns | 123,943,757.93 |
 OrdinalCompareSubstring |       10 | 7.7114 ns | 0.0614 ns | 129,683,834.13 |
  EqualsHelper |       11 | 9.1927 ns | 0.0787 ns | 108,786,813.64 |
 OrdinalCompareSubstring |       11 | 7.8014 ns | 0.2387 ns | 128,261,575.18 |
  EqualsHelper |       12 | 5.9695 ns | 0.0200 ns | 167,518,339.52 |
 OrdinalCompareSubstring |       12 | 5.0780 ns | 0.9026 ns | 200,762,469.14 |
  EqualsHelper |       13 | 5.7687 ns | 0.0493 ns | 173,357,049.75 |
 OrdinalCompareSubstring |       13 | 4.5448 ns | 0.0229 ns | 220,037,374.84 |
  EqualsHelper |       14 | 5.6747 ns | 0.0329 ns | 176,226,182.93 |
 OrdinalCompareSubstring |       14 | 4.6146 ns | 0.0729 ns | 216,739,726.03 |