@benaadams 2016-01-23
`length -= ` can be moved just under the while test rather than last statement in while for better pipelining?