@jkotas 2016-01-23
> the new function does not replace EqualsHelper as it has enough of a different performance profile

I think it is fine for this to be a different method.

BTW: We will apply a profile guided optimizations (aka `IBC`) to the shipping mscorlib. The profile guided optimizations will reorder the method basic blocks based on real-world usage, so it is usually not worth it to optimize the code layout for branch taken vs. branch not taken, etc.