@bbowyersmyth 2016-01-26
Updated with name change. AMD64 => WIN64. length -= moved to the front in EqualsHelper, StartsWithOrdinalHelper, and CompareOrdinalHelper. ALIGN_ACCESS implemented.