@bbowyersmyth 2016-01-26
ALIGN_ACCESS conditional removed. Length checks removed in first int alignment read.